# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SphinX API
  version: 0.1.0
  description: >
    web ↔ server REST 계약. ai-service(내부망 전용)의 /internal/* 인터페이스는 여기 없다 —
    그 경계의 페이로드는 contracts/*.schema.json(judgment·risk_item·parsed_document)으로 계약한다.
paths:
...
  /sessions:
    post:
      summary: 세션 생성 (F-INT-001)
      description: 비식별 속성만 받는다 — 성명·주민번호 필드는 스키마에 존재하지 않는다 (P3).
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateSessionRequest'}
      responses:
        '200':
          description: 생성된 세션(상태 CREATED)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionApiResponse'}
        '400':
          description: 필수 속성 누락 또는 유효성 오류
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}:
    get:
      summary: 세션 상태 조회 (F-INT-001)
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 세션
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionApiResponse'}
        '404':
          description: 세션 없음
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/questions/next:
    post:
      summary: 다음 질문 + 진행 상태 (F-INT-002)
      description: >
        index/total 을 서버가 준다. 화면이 "추출된 항목 수"로 분모를 보완하면 서버가 물어볼
        항목 수와 어긋날 수 있고, 어긋나면 조용히 틀린 진행률이 보인다.
        done=true 가 인터뷰 종료 판단의 유일한 근거다.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 다음 질문(없으면 done=true)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NextQuestionApiResponse'}
  /sessions/{id}/judgments:
    get:
      summary: 세션에 쌓인 항목별 판정 목록 (S-05)
      description: >
        S-03(고객 화면)과 S-05(판매자 화면)는 다른 기기·다른 탭이라 화면이 메모리에 들고
        갈 수 없다. 새로고침에도 살아남아야 한다.
        항목별 signal 은 싣지 않는다 — 게이트 판정은 /judge 의 signal 이 단독 소유한다(P1).
        grade → 색 매핑은 표시 관례이며 판정이 아니다.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 항목 ID 오름차순 판정 목록
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JudgmentsApiResponse'}
        '404':
          description: 세션 없음
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/answers:
    post:
      summary: 텍스트 응답 제출 (F-INT-003) → 항목별 판정(측정값) 반환·세션에 기록
      description: >
        text는 서버에서 PII 마스킹 후 ai-service로 나간다. 반환 Judgment는 '측정값'이며
        게이트 판정이 아니다(P1). 웹 응답은 camelCase(내부 ai-service 계약 judgment.schema.json은 snake_case).
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AnswerRequest'}
      responses:
        '200':
          description: 항목별 이해도 판정
          content:
            application/json:
              schema: {$ref: '#/components/schemas/JudgmentApiResponse'}
        '502':
          description: >
            EVIDENCE_REQUIRED — 상류(ai-service)가 근거 없는 판정을 반환해 차단됐다(P4).
            고객이 고칠 수 있는 입력 오류가 아니므로 400이 아니다. 화면은 재입력을 요구하지
            말고 재시도·황색 보수 처리로 간다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/re-explain:
    post:
      summary: 재설명 (F-INT-004) — 이해 부족 항목을 다시 설명, 이후 같은 항목 재답변이 재검증
      description: >
        재검증 상한(항목당 max-reverify) 도달 항목은 재설명 불가(400) → 판정으로 진행.
        400의 두 경우는 화면 처리가 다르므로 error.code로 구분한다(메시지 문자열은 계약이 아니다).
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReExplainRequest'}
      responses:
        '200':
          description: 재설명 문면(취약 여부로 맞춤)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReExplanationApiResponse'}
        '400':
          description: >
            REEXPLAIN_NOT_ELIGIBLE — 재설명 대상 아님(판정 없음 또는 이미 이해 U1).
            화면은 고객에게 아무것도 띄우지 않고 다음 항목으로 넘어간다.
            REVERIFY_EXHAUSTED — 재검증 상한 도달. 화면은 판정으로 넘어감을 고객에게 알린다
            (F-INT-004 "2회 실패 시 해당 항목 최종 상태 확정 후 게이트로 이관").
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/abort:
    post:
      summary: 세션 중단 (F-INT-001) — 고객 이탈 등 → ABORTED
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 중단된 세션
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SessionApiResponse'}
  /sessions/{id}/simulate:
    post:
      summary: 손실 계산 (F-SIM-001) — 가입금액을 넣으면 시나리오 3건
      description: >
        순수 함수다(P2). 같은 입력이면 항상 같은 출력이며 LLM 이 개입하지 않는다.
        금액은 `data/timeseries/` 실제 종가 시계열을 결정론적 로직에 대입해 산출하고,
        스냅샷은 `data/timeseries/VERSION` 의 sha256 으로 고정된다.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SimulateRequest'}
      responses:
        '200':
          description: 시나리오 3건 (severity 로 worst·mid·best 각 1건)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SimulateApiResponse'}
        '404':
          description: 세션 없음
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/gate-preview:
    get:
      summary: 신호등 미리보기 (F-GTE-001) — 계산만, 기록하지 않음
      description: >
        기획서 7-2 [기능 1] "황색 판정 → 재설명 → 재검증 → 녹색 통과"를 성립시키는 경로다.
        판매자가 황색을 보려고 /judge 를 부르면 JUDGED 로 전이되고 거기서 RE_EXPLAIN 으로
        갈 수 없어 재설명 흐름 자체가 막힌다. 이 경로는 부수효과가 없다.
        GateEngine 이 순수 함수라(P2) 미리보기와 확정이 같은 입력에 같은 답을 낸다.
        이미 판정된 세션은 재계산하지 않고 기록값을 돌려준다(recorded=true).
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 현재 입력 기준 신호등
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GatePreviewApiResponse'}
        '404':
          description: 세션 없음
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
  /sessions/{id}/judge:
    post:
      summary: 게이트 판정 (F-GTE-001) — 세션에 쌓인 판정으로 신호등 산출
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 게이트 판정 결과
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GateApiResponse'}
        '409':
          description: >
            ILLEGAL_STATE_TRANSITION — 판정할 항목이 하나도 없다. JUDGED 는 CLOSE 외에
            나가는 전이가 없어 되돌릴 수 없으므로, 답변 0건 세션이 버튼 오작동 한 번으로
            끝나지 않게 막는다. 신호만 보려면 /gate-preview 를 쓴다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
...
  /sessions/{id}/report:
    post:
      summary: 이해 기록 리포트 발행 (F-GTE-004)
      description: >
        append-only 해시 체인에 기록된 이력에서 리포트를 조립하고 **발행 사실을 체인에 남긴다.**
        발행을 GET 에서 분리한 이유는 감사다 — `report:read` 는 audited action 이라 GET 이
        발행까지 하면 로그에서 "읽었다" 와 "발행했다" 가 구별되지 않고, MGR·COMPL 이 남의
        세션을 열람하는 것만으로 발행 기록이 생긴다. 프리페치·재시도·브라우저 중복 요청이
        상태를 바꾸는 것도 막는다.
        내용이 그대로면 다시 발행하지 않고 기존 것을 돌려준다(멱등). 내용이 달라지면 새로
        발행하되 이전 발행 기록은 지우지 않는다 — "교부 시점에 무엇이 적혀 있었는가" 에
        답하려면 둘이 나란히 남아야 한다(ADR-004).
        권한: `report:read` (SELLER=own_session · MGR=branch · COMPL=org).
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 발행된 리포트(이미 같은 내용이면 기존 것)
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportApiResponse'}
        '403':
          description: >
            범위 밖 접근. `report:read` 는 SELLER=own_session · MGR=branch · COMPL=org 다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
        '404':
          description: 세션 없음
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
    get:
      summary: 발행된 이해 기록 리포트 조회 (F-GTE-004)
      description: >
        **조회만 한다. 상태를 바꾸지 않는다.** 발행한 적 없는 세션은 404 다 — "아직 교부하지
        않았다" 와 "교부했다" 는 감사에서 구별돼야 한다. 발행은 POST 다.
        권한: `report:read`.
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        '200':
          description: 마지막으로 발행된 리포트
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportApiResponse'}
        '403':
          description: 범위 밖 접근
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
        '404':
          description: >
            세션이 없거나 **아직 발행된 리포트가 없다.** 후자는 오류가 아니라 상태다 —
            화면은 "발행하기" 를 띄우면 된다.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorApiResponse'}
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: >
            에러 코드. 프론트는 이 목록을 유니온 타입으로 들고 분기한다 —
            목록에 없는 코드를 서버가 내보내면 화면이 조용히 default로 떨어진다.
          enum:
            - NOT_FOUND                # 404 세션 등 리소스 없음
            - VALIDATION_ERROR         # 400 @Valid 실패
            - MALFORMED_REQUEST        # 400 본문 파싱 실패(잘못된 JSON·허용되지 않은 enum)
            - ILLEGAL_STATE_TRANSITION # 409 허용되지 않은 세션 상태 전이
            - REEXPLAIN_NOT_ELIGIBLE   # 400 재설명 대상 아님(판정 없음 또는 이미 이해 U1)
            - REVERIFY_EXHAUSTED       # 400 재검증 상한 도달 — 판정으로 진행
            - OVERRIDE_NOT_ELIGIBLE    # 409 오버라이드 전제 위반(적색 아님 또는 요청 없이 승인)
            - EVIDENCE_REQUIRED        # 502 P4 위반(근거 없는 판정) — 상류 ai-service 계약 위반
            - AI_SERVICE_UNAVAILABLE   # 502 ai-service 호출 실패(non-2xx·연결 오류·미구현) — 상류 의존성 문제
            - INTERNAL_ERROR           # 500 서버 내부 오류
        message:   {type: string}
        timestamp: {type: string, format: date-time}
    ErrorApiResponse:
      type: object
      required: [success, error]
      properties:
        success: {type: boolean, enum: [false]}
        data:    {type: object, nullable: true}
        error:   {$ref: '#/components/schemas/ApiError'}
    CreateSessionRequest:
      type: object
      description: 세션 생성 입력. 전부 비식별 — 성명·주민번호는 스키마에 존재하지 않는다 (P3).
      required: [productId, channel, ageBand]
      properties:
        productId:       {type: string, description: "상품 ID"}
        channel:         {type: string, enum: [FACE_TO_FACE, MOBILE, TM], description: "판매 채널(대면·모바일·TM)"}
        ageBand:
          type: string
          description: >
            연령대(비식별). 이 값은 server/src/main/resources/vulnerability_weights.yaml 이
            문자열 키로 가중을 매기는 대상이다. 목록에 없는 값을 보내면 400 이 아니라 조용히
            0점이 되고, 취약 고객이 일반 고객으로 분류된다. 그래서 enum 으로 고정한다.
            가중 대상이 아닌 값(20~40대)도 유효한 입력이므로 enum 에 포함한다.
          enum: ["20대", "30대", "40대", "50대", "60대", "70대", "80대"]
        experienceLevel:
          type: string
          nullable: true
          description: 투자 경험 수준. ageBand 와 같은 이유로 enum 고정.
          enum: ["없음", "1년미만", "3년미만", "3년이상"]
        amountBand:
          type: string
          nullable: true
          description: 가입금액대(비식별). ageBand 와 같은 이유로 enum 고정.
          enum: ["1천만원미만", "1천만원대", "3천만원대", "5천만원대", "1억원대", "3억원대이상"]
        contractRef:     {type: string, nullable: true, description: "계약건 참조번호(비식별)"}
        surveySchemaVersion: {type: string, nullable: true, description: "적합성 설문 문항 세트 버전 — 리포트가 어느 세트로 받았는지 판별(F-GTE-004)"}
        surveyResult:    {type: object, nullable: true, additionalProperties: true, description: "적합성 설문 결과(문항 식별자→답변)"}
    # data 의 모양을 소유자가 아직 정하지 않은 응답용 봉투. 봉투만 계약하고 안은 열어 둔다.
    # 프론트는 success 로 분기하고 data 는 해당 화면이 아는 형태로 읽는다.
...
    NextQuestion:
      type: object
      required: [index, total, done]
      properties:
        itemId:   {type: string, nullable: true, description: "done=true 면 null"}
        question: {type: string, nullable: true, description: "done=true 면 null"}
        index:    {type: integer, description: "1-based. done=true 면 total 과 같다"}
        total:    {type: integer, description: "서버가 물어볼 항목 수. 추출 항목 수와 다를 수 있다"}
        done:     {type: boolean, description: "더 물을 항목이 없다. 인터뷰 종료 판단의 유일한 근거"}
    NextQuestionApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/NextQuestion'}
        error:   {type: object, nullable: true}
    JudgmentsResponse:
      type: object
      required: [sessionId, state, judgments]
      properties:
        sessionId: {type: string}
        state:     {type: string, enum: [CREATED, IN_PROGRESS, RE_EXPLAIN, RE_VERIFY, JUDGED, CLOSED, ABORTED]}
        judgments:
          type: array
          description: 항목 ID 오름차순. 항목별 최신 판정만 담긴다(재검증 시 덮어씀)
          items: {$ref: '#/components/schemas/JudgmentView'}
    JudgmentsApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/JudgmentsResponse'}
        error:   {type: object, nullable: true}
    SessionResponse:
      type: object
      required: [sessionId, state, productId]
      properties:
        sessionId:   {type: string}
        state:       {type: string, enum: [CREATED, IN_PROGRESS, RE_EXPLAIN, RE_VERIFY, JUDGED, CLOSED, ABORTED]}
        productId:   {type: string}
        contractRef: {type: string, nullable: true}
    SessionApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/SessionResponse'}
        error:   {type: object, nullable: true}
    AnswerRequest:
      type: object
      description: F-INT-003 텍스트 응답. text는 서버에서 PII 마스킹 후 ai-service로.
      required: [itemId, text]
      properties:
        itemId:    {type: string}
        text:      {type: string, description: "고객 발화(서버에서 마스킹)"}
        inputMeta: {type: object, nullable: true, additionalProperties: true, description: "붙여넣기·지연·수정빈도 등(F-INT-003)"}

    # 웹 응답용 Judgment(camelCase). 내부 ai-service 계약은 judgment.schema.json(snake_case).
    JudgmentView:
      type: object
      required: [itemId, grade, confidence, evidence, reason]
      properties:
        itemId:     {type: string}
        grade:      {type: string, enum: [U1, U2, U3, U4]}
        confidence: {type: number, minimum: 0, maximum: 1}
        evidence:
          type: object
          required: [utteranceQuote, rubricClause]
          properties:
            utteranceQuote: {type: string}
            rubricClause:   {type: string}
        reason:            {type: string}
        misconceptionType: {type: string, nullable: true}
    JudgmentApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/JudgmentView'}
        error:   {type: object, nullable: true}
    ReExplainRequest:
      type: object
      required: [itemId]
      properties:
        itemId: {type: string, description: "다시 설명할 이해항목"}
    ReExplanation:
      type: object
      required: [itemId, content, vulnerable, reverifyQuestion]
      properties:
        itemId:     {type: string}
        content:    {type: string, description: "재설명 문면(취약 시 고령자 모드)"}
        vulnerable:
          type: boolean
          description: >
            취약 고객 여부. 렌더링 힌트(큰 글씨·비유)이지 화면에 표시할 라벨이 아니다 —
            "취약 고객으로 분류됨"을 본인에게 보이지 않는다(기획서 7-4, 보호 목적으로만).
        reverifyQuestion:
          type: string
          description: >
            재설명 직후 같은 항목을 다시 물을 때 쓸 변형 질문(F-INT-002). 프론트는 직전 질문을
            재사용하지 말고 이 값을 띄운다. 서버 구현이 목인 동안에는 항목별 고정 문면이라
            기획서 7-4 1단계(우회 비용 상향)를 아직 만족하지 않는다 — 문서 자동 생성이
            붙어야 성립한다. 계약 형태는 그때도 바뀌지 않는다.
    ReExplanationApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/ReExplanation'}
        error:   {type: object, nullable: true}
    GatePreview:
      type: object
      required: [signal, ruleTrace, recorded]
      properties:
        signal:    {type: string, enum: [GREEN, YELLOW, RED]}
        ruleTrace: {type: array, items: {type: string}, description: "발화한 룰 ID(예: R-01)"}
        recorded:
          type: boolean
          description: >
            감사 기준점으로 기록된 값인가. false 면 아직 확정이 아니다 — 화면이 이걸
            확정으로 보관하면 안 된다. /judge 를 부른 뒤에만 true 가 된다.
        judgedAt:
          type: string
          format: date-time
          nullable: true
          description: "기준점 확정 시각. null 이면 /judge 가 아직 호출되지 않은 세션이다"
    GatePreviewApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/GatePreview'}
        error:   {type: object, nullable: true}
    # ── F-SIM-001 손실 시뮬레이터 (소유: 정세현) ──────────────────────────────
    SimScenario:
      type: object
      required: [severity, result, name, payout, pnl, pathMeta]
      properties:
        severity:
          type: string
          enum: [worst, mid, best]
          description: >
            화면 배치·정렬 키다. **표시 라벨이 아니다** — 기획서 7-2 표는 평가어
            ("최선/중간/최악")를 버리고 금액 순으로 갔다. 그 이유는 스텝다운이 조기상환
            시점과 무관하게 연 수익률이 같고 총액만 달라서, 가장 자주 일어나는 전개가
            금액으로는 가장 작기 때문이다("최선이 왜 더 작은가"라는 잘못된 질문이 생긴다).
            화면은 명세 8절 S-04 의 3열 동일 비중 요건을 위해 이 키로 자리를 고정하고,
            사람에게 보일 문면은 `name` 을 쓴다.
        result:
          type: string
          description: "정확한 상환 유형. early_1..early_5 | maturity | loss (SimulatorService.Outcome.result 그대로)"
        name:   {type: string, description: "표시 문면(예: 낙인 45% 하회 후 만기 손실)"}
        payout: {type: integer, format: int64, description: "상환금액(원). 원 단위 절사"}
        pnl:    {type: integer, format: int64, description: "손익(원). 음수는 손실"}
        share:  {type: number, nullable: true, minimum: 0, maximum: 1, description: "역사 전 구간에서 이 전개가 발생한 비중"}
        pathMeta: {$ref: '#/components/schemas/PathMeta'}
    PathMeta:
      type: object
      description: >
        어느 실제 지수 구간을 썼는지. 명세 8절 S-04 의 "역사 구간 라벨" 을 그리는 데 쓰고,
        동시에 P2 재현성의 실물 근거가 된다 — 화면에 스냅샷을 박으면 "이 수치는 이
        스냅샷에서 나왔다"가 심사자에게 보인다.
      required: [startDate, endDate, underlyings, worstUnderlying, worstFinal, knockedIn]
      properties:
        startDate:       {type: string, format: date, description: "계약일"}
        endDate:         {type: string, format: date, description: "상환일(조기상환 또는 만기)"}
        underlyings:     {type: array, items: {type: string}, description: "기초자산 키(sp500·nikkei225·eurostoxx50)"}
        worstUnderlying: {type: string, description: "만기평가일 최저 종목 — 손실 판정의 기준"}
        worstFinal:      {type: number, description: "그 종목의 최초기준가격 대비 비율(예: 0.507)"}
        knockedIn:       {type: boolean, description: "낙인 배리어를 하회한 적이 있는가"}
    SimulateRequest:
      type: object
      description: >
        `amount` 에 서버 기본값을 두지 않는다. 금액은 화면이 슬라이더로 정하는 값이라
        서버가 임의 기본값을 갖고 있으면 프론트 버그가 조용히 그럴듯한 숫자로 덮인다.
      required: [amount]
      properties:
        amount: {type: integer, format: int64, minimum: 1, description: "가입금액(원, 양수)"}
    SimulateResponse:
      type: object
      required: [scenarios, timeseriesVersion, productName]
      properties:
        scenarios:
          type: array
          minItems: 3
          maxItems: 3
          items: {$ref: '#/components/schemas/SimScenario'}
          description: "severity 별로 정확히 1건씩"
        timeseriesVersion: {type: string, description: "data/timeseries/VERSION 의 snapshot(예: 2026-08-24). 화면에 표시한다"}
        productName:       {type: string, description: "상품 조건의 출처(예: 키움증권 제4181회 ELS)"}
    SimulateApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/SimulateResponse'}
        error:   {type: object, nullable: true}

    # ── F-GTE-004 이해 기록 리포트 (소유: 정세현) ─────────────────────────────
    ReportResponse:
      type: object
      required: [reportId, sessionId, generatedAt, contentHash]
      properties:
        reportId:    {type: string}
        sessionId:   {type: string}
        generatedAt: {type: string, format: date-time}
        contentHash:
          type: string
          description: >
            리포트 **내용**을 CanonicalJson 으로 직렬화한 sha256. 체인의 항목 해시가 아니다 —
            항목 해시는 앞 항목과 순번에 의존하므로 문서 한 장을 든 고객이 재계산할 수 없고,
            그러면 해시를 싣는 이유가 사라진다. 요약본과 같은 값이어야 하는 이유이기도 하다.
            발행 시각은 내용에 넣지 않는다 — 넣으면 같은 내용을 두 번 발행할 때마다 해시가
            달라져 "이 문서가 그 문서인가" 를 대조할 수 없다.
            검증은 두 층이다. (1) 체인이 온전한가는 `audit:verify`. (2) 이 문서가 그 기록에서
            나왔는가는 기록을 다시 조립해 이 해시와 대조한다. 발행 사실 자체는 체인에 남고
            그 항목의 payload 가 이 값을 담는다.
        previewUrl:
          type: string
          nullable: true
          description: >
            브라우저 미리보기(PDF 인라인). **PDF 생성이 붙기 전까지 null 이다.** 값을 채우면
            계약이 "이 URL 로 가면 문서가 있다" 를 보장하는데 404 가 난다 — 스키마 검증은
            통과하고 화면은 링크를 그리며, 눌러야 드러난다.
        downloadUrl:
          type: string
          nullable: true
          description: "PDF 다운로드. previewUrl 과 같은 이유로 PDF 생성 전까지 null."
    ReportApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/ReportResponse'}
        error:   {type: object, nullable: true}
...
    GateResult:
      type: object
      required: [signal, ruleTrace]
      properties:
        signal:    {type: string, enum: [GREEN, YELLOW, RED]}
        ruleTrace: {type: array, items: {type: string}, description: "발화한 룰 ID(예: R-01)"}
    GateApiResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:    {$ref: '#/components/schemas/GateResult'}
        error:   {type: object, nullable: true}